==> src/data-preparation/3_data_exploration.Rmd <==
---
title: "Data Exploration"
subtitle: "SKILLS: Data Preparation and Workflow Management - Group 9"
author:
  - "Rabino Tommaso"
  - "Franceschini Emanuele"
  - "Magalotti Bianca"
  - "Tan Colin"
  - "Benmrit Akram"
date: "\\textit{20 October 2024}"
output:
  pdf_document:
    latex_engine: xelatex
    toc: true
    toc_depth: 2
    number_sections: true
    fig_caption: true
header-includes:
  - "\\usepackage[english]{babel}"  # Language
  - "\\usepackage[T1]{fontenc}"  # Font encoding
  - "\\usepackage{mathptmx}"  # Times New Roman font for text
  - "\\usepackage{helvet}"  # Arial-like font for sans-serif
  - "\\usepackage{setspace}"  # Line spacing
  - "\\onehalfspacing"  # 1.5 line spacing
  - "\\usepackage{fancyhdr}"  # Header and footer customization
  - "\\usepackage{titlesec}"  # Section titles formatting
  - "\\usepackage{abstract}"  # Abstract formatting
  - "\\usepackage{caption}"  # Captions customization
  - "\\usepackage{graphicx}"  # Graphics
  - "\\usepackage{amsmath}"  # Math equations
  - "\\usepackage{amssymb}"  # Math symbols
  - "\\usepackage{natbib}"  # Citation style (change as needed)
  - "\\bibliographystyle{apalike}"  # Bibliography style (change as needed)
  - "\\usepackage{hyperref}"  # Hyperlinks and URLs
  - "\\hypersetup{colorlinks=true,linkcolor=blue,citecolor=blue,urlcolor=blue}"
  - "\\usepackage{appendix}"  # Appendix formatting
  - "\\usepackage{enumerate}"  # Enumerate environment
  - "\\pagestyle{fancy}"  # Custom page style
  - "\\fancyhf{}"
  - "\\renewcommand{\\headrulewidth}{0pt}"
  - "\\renewcommand{\\footrulewidth}{0pt}"
  - "\\fancyhead[R]{\\thepage}"  # Page number in the header
  - "\\fancypagestyle{plain}{\\fancyhf{}\\renewcommand{\\headrulewidth}{0pt}}"
  - "\\lhead{\\small{A.A. 2023/2024-Courses: SKILLS: Data Preparation and Workflow Management}}"
  - "\\usepackage{multicol}"  # For two columns
geometry: "left=2.5cm, right=1.5cm, top=2.5cm, bottom=2.5cm"  # Adjust margins
---

```{r, include=FALSE}
knitr::opts_chunk$set(tidy = FALSE,
                      size = "small",
                      out.width = "100%",
                      message = FALSE,
                      warning = FALSE,
                      error = FALSE)
```

```{r, include=FALSE}
#Setting the seed for reproducible results
set.seed(999)
```

```{r, include = FALSE}
###############################################################
#############             PACKAGES               ##############
###############################################################
#GENERAL PACKAGES:
library(readr)
library(tidyverse) #A "Package of Packages" for Data manipulation and visualization (includes magrittr, lubridate, purrr, tidyr, etc.).
library(dplyr) #Data frame manipulations (select, slice, etc.
library(jsonlite) #For Amenities Columns Creation
library(moments) #Measuring the skewness.

#REGRESSION PACKAGES
library(caret) #Hyperparameters Tuning. 
library(xgboost) #XGBoost Regression. 
library(DALEX) #Summary of the XGBoost Regression Model ("explainer).
library(bayesforecast) #Checking Regression Assumptions.

#SHINYAPP PACKAGES
library(shiny) #For the ShinyApp
library(shinyWidgets) #For the ShinyApp

#PLOT AND FIGURES PACKAGES:
library(ggplot2) #Building fancy plots.
library(ggthemes) #Themes for ggplots (e.g. "solarized").
library(ggcorrplot) #For correlograms
library(scales) #Scaling and formatting ggplots (e.g. scale_fill_gradient()).
library(gt) #Latex tables

# WORKING DIRECTORY SETTING PACKAGES
library(here)
library(rstudioapi)
```

```{r, include=FALSE}
###############################################################
######      AUTOMATICALLY SET WORKING DIRECTORY         #######
###############################################################

if (rstudioapi::isAvailable()) {
  # Get the path of the active document
  current_document_path <- getActiveDocumentContext()$path
  
  # Check if the document path is set
  if (!is.null(current_document_path)) {
    # Set the working directory to the source file's location
    setwd(dirname(current_document_path))
    
    # Optionally, you can remove the document path variable
    rm(current_document_path)
  } else {
    cat("No active document or path found. Set working directory manually.\n")
  }
} else {
  # This section is executed when not in RStudio
  # Here, you can set the working directory manually or leave it at its default
  cat("Script is not running in RStudio. Consider setting the working directory manually if required.\n")
}
```

```{r, echo=FALSE}
# Load the R object
clean_data <- readRDS("../../gen/data-preparation/input/clean_data.rds")
```


\newpage
\section{Correlation Analysis}
This section is structured as follows:

\begin{enumerate}
  \item Correlation Analysis of Apartment-related Variables: In this step, a correlogram is generated to examine the correlations between \textit{numeric}, \textit{apartment-related} variables and the price of the listings.
  \item Correlation Analysis of Host-related Variables: Subsequently, a correlogram is constructed to assess the correlations between \textit{numeric}, \textit{host-related} variables and the price of the listings.
\end{enumerate}


\subsection{Apartment-related Variables}
```{r, include=FALSE}
# Select relevant variables for correlation analysis
corr_data <- clean_data %>% 
  dplyr::select(
    price, latitude, longitude, accommodates, bedrooms, beds, bathrooms, minimum_nights,
    maximum_nights, availability_30, availability_60, availability_90, availability_365,
    number_of_reviews, number_of_reviews_ltm, number_of_reviews_l30d, reviews_per_month,
    review_scores_rating, review_scores_accuracy, review_scores_cleanliness, review_scores_location,
    review_scores_value
  )

```

Here is a correlogram displaying correlation values for \texttt{numeric} \textit{apartment-related} variables:
```{r, include=FALSE}
# Create a correlation matrix
corr_matrix <- round(cor(corr_data),2)

# Define custom color palette
my_colors <- c("#E6E6E6", "#FFFFFF", "#CCCCCC")

# Restyled correlogram plot
plot <- ggcorrplot(corr_matrix,
           hc.order = FALSE,
           method = "square",
           type = "lower",
           outline.col = "white",
           colors = my_colors,
           lab = TRUE, 
           lab_size = 2.5,
           digits = 2,
           tl.cex = 8,
           title = "Apartment-related Variables Correlogram") +
  theme(legend.position = "none", 
        plot.title = element_text(lineheight = 2, face = "bold", vjust = 1, hjust = 0.5, color = "#333333"),
        plot.subtitle = element_text(color = "#666666", vjust = 1, hjust = 0.5),
        axis.title = element_text(color = "#666666"), 
        axis.text = element_text(color = "#666666"))

```

```{r, echo=FALSE}
plot
```


To improve clarity, we created a table displaying the correlation of each variable with \texttt{price}. Additionally, we've organized the variables in descending order based on their correlation with \texttt{price} to enhance interpretation.

The table below reveals that certain variables exhibit a stronger correlation with \texttt{price}. Notably, \texttt{bathrooms}	(0.44), \texttt{bedrooms}	(0.42), \texttt{accommodates} (0.39), and \texttt{beds} (0.32), show the highest positive correlations.
Conversely, all \texttt{review\_score} related variables have weak, or even negative, correlations with \texttt{price}.

```{r, include=FALSE}
# Convert correlation matrix into a dataframe
corr_matrix_df <- data.frame(corr_matrix)

# Add a new column for variable names
corr_matrix_df$Variable <- row.names(corr_matrix_df)

# Select the "Variable" and "price" columns
corr_matrix_df <- corr_matrix_df %>%
  select(Variable, price) %>%
  arrange(desc(price))

# Rename the columns
colnames(corr_matrix_df) <- c("Variable", "Correlation")


table_corr_price_apartment <- corr_matrix_df %>%
  gt() %>%
  tab_style(
    style = cell_text(weight = "bold"),
    locations = cells_column_labels(columns = everything())
  ) %>%
  tab_header(
    title = "Correlation with Price",
    subtitle = "Correlation between Variables and Price"
  )
```

```{r, echo=FALSE}
table_corr_price_apartment
```




\subsection{Host-related Variables}
Here is a correlogram displaying correlation values for \texttt{numeric} \textit{host-related} variables:
```{r, include=FALSE}
# Select relevant variables
corr_data <- clean_data %>% 
  dplyr::select(price, host_response_rate_percentage, host_acceptance_rate_percentage, host_listings_count, host_total_listings_count, review_scores_checkin, review_scores_communication, reviews_per_month)

corr_matrix <- round(cor(corr_data),2)

# Define custom color palette
my_colors <- c("#E6E6E6", "#FFFFFF", "#CCCCCC")

# Restyled correlogram plot
plot <- ggcorrplot(corr_matrix,
           hc.order = FALSE,
           method = "square",
           type = "lower",
           outline.col = "white",
           colors = my_colors,
           lab = TRUE, 
           lab_size = 2.5,
           digits = 2,
           tl.cex = 8,
           title = "Host-related Variables Correlogram") +
  theme(legend.position = "none", 
        plot.title = element_text(lineheight = 2, face = "bold", vjust = 1, hjust = 0.5, color = "#333333"),
        plot.subtitle = element_text(color = "#666666", vjust = 1, hjust = 0.5),
        axis.title = element_text(color = "#666666"), 
        axis.text = element_text(color = "#666666"))

```

```{r, echo=FALSE}
plot
```


To improve clarity, we created a table displaying the correlation of each variable with \texttt{price}. Additionally, we've organized the variables in descending order based on their correlation with \texttt{price} to enhance interpretation.

The table below reveals that certain variables exhibit a stronger correlation with \texttt{price}. Notably, \texttt{host\_listings\_count} (0.07), and \texttt{host\_total\_listings\_count} (0.06), show the highest positive correlations. Conversely, all \texttt{review\_score} related variables have weak, or even negative, correlations with \texttt{price}. Moreover, in general, \texttt{host-related} variables show weaker correlations with \texttt{price} if compared to \texttt{apartment-related} variables.

```{r, include=FALSE}
# Convert correlation matrix into a dataframe
corr_matrix_df <- data.frame(corr_matrix)

# Add a new column for variable names
corr_matrix_df$Variable <- row.names(corr_matrix_df)

# Select the "Variable" and "price" columns
corr_matrix_df <- corr_matrix_df %>%
  select(Variable, price) %>%
  arrange(desc(price))

# Rename the columns
colnames(corr_matrix_df) <- c("Variable", "Correlation")


table_corr_price_host <- corr_matrix_df %>%
  gt() %>%
  tab_style(
    style = cell_text(weight = "bold"),
    locations = cells_column_labels(columns = everything())
  ) %>%
  tab_header(
    title = "Correlation with Price",
    subtitle = "Correlation between Variables and Price"
  )
```

```{r, echo=FALSE}
table_corr_price_host
```


\section{Amenities-related Variables}
This code chunk iterates through all the \texttt{amenities-related} variables within the dataframe and computes the median price difference between listings that offer a specific amenity and those that do not.
```{r}
# Create an empty dataframe to store the results
median_diff_df <- data.frame(Variable_Name = character(0), Median_Difference = numeric(0))

# Loop through variables 35 to 127
for (i in 42:123) {
  # Select the relevant columns and group by Bidet
  amenities_df <- clean_data %>%
    select(price, all_of(i)) %>%
    group_by(across(all_of(names(.)[2]))) %>%
    summarize(median_price = median(price))
  
  # Calculate the median difference
  median_diff <- amenities_df$median_price[2] - amenities_df$median_price[1]
  
  # Append the result to the dataframe
  median_diff_df <- rbind(median_diff_df, data.frame(Variable_Name = colnames(clean_data)[i], Median_Difference = median_diff))
}
```

The following table displays the top ten median price differences among all amentity. 
```{r, include=FALSE}
# Select the first 10 rows and arrange by Median_Difference
top_10_median_diff <- median_diff_df %>%
  arrange(desc(Median_Difference)) %>%
  head(10)

# Append "€" symbol to the second column of your dataset
top_10_median_diff$Median_Difference <- paste(top_10_median_diff$Median_Difference, "€")

# Create a gt table
table_higher_diff_amenities <- top_10_median_diff %>%
  gt() %>%
  tab_style(
    style = cell_text(weight = "bold"),
    locations = cells_column_labels(columns = everything())
  ) %>%
  tab_header(
    title = "Top 10 Amenities with Highest Median Price Difference",
    subtitle = "Based on Median Price Difference",
  )

```

```{r, echo=FALSE}
table_higher_diff_amenities
```



\section{Property\_type variable}
\subsection{Top 10 Most Expensive Property Types}
The following table displays the top 10 most costly property types.

```{r, include=FALSE}

property_type_grouped <- clean_data%>%
  dplyr::group_by(property_type)%>%
  dplyr::summarize(avg_price_property_type = round(mean(price), 2))

property_type_grouped <- property_type_grouped%>%
  dplyr::arrange(desc(avg_price_property_type))%>%
  dplyr::slice(1:10)

table_expensive_property_type <- property_type_grouped %>%
  gt() %>%
  tab_header(
    title = "Average Price of Listings per Property Type",
    subtitle = "Top 10 Most Expensive Property Types"
  )

```


```{r, echo=FALSE}
table_expensive_property_type
```


\subsection{Top 10 Cheapest Property Type}
The following table displays the top 10 most cheapest property types.

```{r, include=FALSE}
property_type_grouped <- clean_data%>%
  dplyr::group_by(property_type)%>%
  dplyr::summarize(avg_price_property_type = round(mean(price), 2))

property_type_grouped <- property_type_grouped%>%
  dplyr::arrange(avg_price_property_type)%>%
  dplyr::slice(1:10)%>%
  dplyr::arrange(desc(avg_price_property_type))


table_cheap_property_type <- property_type_grouped %>%
  gt() %>%
  tab_header(
    title = "Average Price of Listings per Property Type",
    subtitle = "Top 10 Cheapest Property Types"
  )


```


```{r, echo=FALSE}
table_cheap_property_type
```


\section{Neighbourhood\_cleansed variable}
\subsection{Top 10 Most Expensive Neighbourhoods}
The table below displays the top 10 most costly neighbourhoods.

```{r, include=FALSE}

neighbourhood_grouped <- clean_data%>%
  dplyr::group_by(neighbourhood_cleansed)%>%
  dplyr::summarize(avg_price_neighbourhood = round(mean(price), 2))

neighbourhood_grouped <- neighbourhood_grouped%>%
  dplyr::arrange(desc(avg_price_neighbourhood))%>%
  dplyr::slice(1:10)

table_expensive_neighbourhood <- neighbourhood_grouped %>%
  gt() %>%
  tab_header(
    title = "Average Price of Listings per Neighbourhood",
    subtitle = "Top 10 Most Expensive Neighbourhoods"
  )
```


```{r, echo=FALSE}
table_expensive_neighbourhood
```


\subsection{Top 10 Cheapest Neighbourhoods}
The table below displays the top 10 most cheapest neighbourhoods.

```{r, include=FALSE}
neighbourhood_grouped <- clean_data%>%
  dplyr::group_by(neighbourhood_cleansed)%>%
  dplyr::summarize(avg_price_neighbourhood = round(mean(price), 2))

neighbourhood_grouped <- neighbourhood_grouped%>%
  dplyr::arrange(avg_price_neighbourhood)%>%
  dplyr::slice(1:10)%>%
  dplyr::arrange(desc(avg_price_neighbourhood))
  

table_cheap_neighbourhood <- neighbourhood_grouped %>%
  gt() %>%
  tab_header(
    title = "Average Price of Listings per Neighbourhood",
    subtitle = "Top 10 Cheapest Neighbourhoods"
  )

```


```{r, echo=FALSE}
table_cheap_neighbourhood
```




```{r, include=FALSE}
rm(list=ls())
```
